pyproject.toml
[project]
name = "llm-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.69.0",
    "beautifulsoup4>=4.14.2",
    "chromadb>=1.1.1",
    "datasets==3.6.0",
    "feedparser>=6.0.12",
    "gensim>=4.3.2",
    "google-generativeai>=0.8.5",
    "gradio>=5.49.0",
    "ipywidgets>=8.1.7",
    "jupyter-dash>=0.4.2",
    "jupyterlab>=4.4.9",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.78",
    "langchain-openai>=0.3.34",
    "langchain-text-splitters>=0.3.11",
    "matplotlib>=3.10.6",
    "modal>=1.1.4",
    "notebook>=7.4.7",
    "numpy>=1.26.4",
    "ollama>=0.6.0",
    "openai>=2.1.0",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "protobuf==3.20.2",
    "psutil>=7.1.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.13.1",
    "sentence-transformers>=5.1.1",
    "setuptools>=80.9.0",
    "speedtest-cli>=2.1.3",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.0",
    "wandb>=0.22.1",
]
